# Reduce allocations in `PathVar`