--- a/dsl/src/main/scala/org/http4s/dsl/impl/Path.scala
+++ b/dsl/src/main/scala/org/http4s/dsl/impl/Path.scala
@@ -36,8 +36,6 @@ import org.http4s.Uri.Path._
 import org.http4s._
 import org.http4s.headers.Allow
 
-import scala.util.Try
-
 object :? {
   def unapply[F[_]](req: Request[F]): Some[(Request[F], Map[String, collection.Seq[String]])] =
     Some((req, req.multiParams))
@@ -166,10 +164,10 @@ object /: {
     }
 }
 
-protected class PathVar[A](cast: String => Try[A]) {
+protected class PathVar[A](cast: String => Option[A]) {
   def unapply(str: String): Option[A] =
-    if (!str.isEmpty)
-      cast(str).toOption
+    if (str.nonEmpty)
+      cast(str)
     else
       None
 }
@@ -180,7 +178,14 @@ protected class PathVar[A](cast: String => Try[A]) {
   *      case Root / "user" / IntVar(userId) => ...
   * }}}
   */
-object IntVar extends PathVar(str => Try(str.toInt))
+object IntVar
+    extends PathVar((str: String) =>
+      try
+        Option(str.toInt)
+      catch {
+        case t if scala.util.control.NonFatal(t) => None
+      }
+    )
 
 /** Long extractor of a path variable:
   * {{{
@@ -188,7 +193,14 @@ object IntVar extends PathVar(str => Try(str.toInt))
   *      case Root / "user" / LongVar(userId) => ...
   * }}}
   */
-object LongVar extends PathVar(str => Try(str.toLong))
+object LongVar
+    extends PathVar((str: String) =>
+      try
+        Option(str.toLong)
+      catch {
+        case t if scala.util.control.NonFatal(t) => None
+      }
+    )
 
 /** UUID extractor of a path variable:
   * {{{
@@ -196,7 +208,14 @@ object LongVar extends PathVar(str => Try(str.toLong))
   *      case Root / "user" / UUIDVar(userId) => ...
   * }}}
   */
-object UUIDVar extends PathVar(str => Try(java.util.UUID.fromString(str)))
+object UUIDVar
+    extends PathVar((str: String) =>
+      try
+        Option(java.util.UUID.fromString(str))
+      catch {
+        case t if scala.util.control.NonFatal(t) => None
+      }
+    )
 
 /** Matrix path variable extractor
   * For an example see [[https://www.w3.org/DesignIssues/MatrixURIs.html MatrixURIs]]