Before:
```
[info] PathVarBench.intVarPath                                   avgt   10  5604.150 ± 482.813   ns/op
[info] PathVarBench.intVarPath:·gc.alloc.rate.norm               avgt   10  4857.237 ±   0.024    B/op
[info] PathVarBench.intVarPath:·gc.churn.G1_Eden_Space.norm      avgt   10  4927.337 ±  45.360    B/op
[info] PathVarBench.intVarPath:·gc.churn.G1_Survivor_Space.norm  avgt   10     0.012 ±   0.008    B/op
```
After:
```
[info] PathVarBench.intVarPath                                   avgt   10  5178.997 ± 48.673   ns/op
[info] PathVarBench.intVarPath:·gc.alloc.rate.norm               avgt   10  4681.195 ±  0.013    B/op
[info] PathVarBench.intVarPath:·gc.churn.G1_Eden_Space.norm      avgt   10  4743.538 ± 29.411    B/op
[info] PathVarBench.intVarPath:·gc.churn.G1_Survivor_Space.norm  avgt   10     0.010 ±  0.004    B/op
```
How it was benchmarked:
```scala
@BenchmarkMode(Array(Mode.AverageTime))
@OutputTimeUnit(TimeUnit.NANOSECONDS)
@State(Scope.Benchmark)
class PathVarBench {
  val intVarPathHttpRoutes =
    HttpRoutes.of[IO] {
      case GET -> Root / "foo" / IntVar(i) =>
        Ok(i.toString)
    }
  val intVarPathRequest = Request[IO](Method.GET, uri"/foo/42")

  @Benchmark
  def intVarPath =
    intVarPathHttpRoutes(intVarPathRequest).value.unsafeRunSync()
}
```